Fix security concerns about connectionString
Previously connectionString was saved in appsettings.json file, which is considered an unsafe practice
as there's no level of protection. Such a file can be read easily on a webserver or be leaked with source
code.
Therefore an alternative like environment variables should be tried instead. Environment variables offer
more security as they're not part of source code and reduce the risk of accidental leaks.