Route error handling with new components
Errors now can be displayed as toasts or have their own dedicated components.
Differentiating between them offers a better interface than simple toasts as
a Not Found error is generic and can be applied in case of a missing resource while a server error
is intended for a more readable form in development.