Optimize loading speed for items in cart
Before adding items to basket would be reflected a bit late into
the basket icon inside navbar or UI.
Now with an optimistic update this problem is handled inside the
mutating queries.Cache invalidation is removed and replaced by operations
that modify the basket in the browser without waiting for a server
request. If the server request fails the changes are rolled back.